feat: implement nineball practice mode and refactor tests
- Implement Nine Ball practice mode (any ball 1-8 in any order).
- Trigger via 'practice' URL parameter.
- Centralize practice state in Session singleton.
- Update NineBall rules (foulReason, isEndOfGame, handlePot) for practice.
- Consolidate practice mode tests into test/rules/nineball.spec.ts to resolve CI duplication failure.
- Add practice.md analysis document.
